label

+++ iDISCO/Analysis/Voxelization.py
@@ -29,7 +29,7 @@ def voxelize(points, imagesize, average = (5,5,5), mode = 'Spherical'):
 
 
 def voxelizePixel(points, imagesize):
-    vox = numpy.zeros(imagesize, dtype=numpy.int32);
+    vox = numpy.zeros(imagesize, dtype=numpy.int16);
     for i in range(points.shape[0]):
         if points[i,0] > 0 and points[i,0] < imagesize[0] and points[i,1] > 0 and points[i,1] < imagesize[1] and points[i,2] > 0 and points[i,2] < imagesize[2]:
                vox[points[i,0], points[i,1], points[i,2]] = 1;
